% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_best.R
\name{rank_best}
\alias{rank_best}
\title{Rank Best Optimal Weights}
\usage{
rank_best(table, bounds = c(0, 5), cleanup = TRUE)
}
\arguments{
\item{table}{Table of indicators}

\item{bounds}{Bounds for integer weights (default 0-5)}

\item{cleanup}{Flag to remove .cip and .sol files (default TRUE)}
}
\value{
A table of weights in order of competitors
}
\description{
Rank Best Optimal Weights
}
